New feature allowing to disable sending FT to smart contracts (#684)
# Description

There are 4 entry points now, which are handled to detect if sender or recipient is a smart contract:
- single send in bank keeper
- multi send in bank keeper
- handler of messages executed by smart contract
- smart contract instantiation

In those entry points, addresses are added to appropriate list in context (senders or recipients) and then interested parties may verify if sender or recipient exists in the list, meaning that this is smart contract.

It might happen that one address is handled by two entry points. For example, if previously instantiated smart contract is executed with coins attached, the address of smart contract will be marked in two places:
- WASM token transferrer
- bank keeper

In this PR, smart contract instantiation is rejected if its address already exists. Otherwise, funds could be sent there in advance, before contract instantiation.

# Reviewers checklist:
- [ ] Try to write more meaningful comments with clear actions to be taken.
- [ ] Nit-picking should be unblocking. Focus on core issues.

# Authors checklist
- [x] Provide a concise and meaningful description
- [x] Review the code yourself first, before making the PR.
- [x] Annotate your PR in places that require explanation.
- [x] Think and try to split the PR to smaller PR if it is big.